// apps/webapp/spa/src/lib/api/generated/models/BrandArchetypeName.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Voz AI API
 * Voz AI marketing automation platform API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const BrandArchetypeName = {
    innocent: 'INNOCENT',
    everyman: 'EVERYMAN',
    hero: 'HERO',
    outlaw: 'OUTLAW',
    explorer: 'EXPLORER',
    creator: 'CREATOR',
    ruler: 'RULER',
    magician: 'MAGICIAN',
    lover: 'LOVER',
    caregiver: 'CAREGIVER',
    jester: 'JESTER',
    sage: 'SAGE'
} as const;
export type BrandArchetypeName = typeof BrandArchetypeName[keyof typeof BrandArchetypeName];


export function instanceOfBrandArchetypeName(value: any): boolean {
    for (const key in BrandArchetypeName) {
        if (Object.prototype.hasOwnProperty.call(BrandArchetypeName, key)) {
            if (BrandArchetypeName[key as keyof typeof BrandArchetypeName] === value) {
                return true;
            }
        }
    }
    return false;
}

export function BrandArchetypeNameFromJSON(json: any): BrandArchetypeName {
    return BrandArchetypeNameFromJSONTyped(json, false);
}

export function BrandArchetypeNameFromJSONTyped(json: any, ignoreDiscriminator: boolean): BrandArchetypeName {
    return json as BrandArchetypeName;
}

export function BrandArchetypeNameToJSON(value?: BrandArchetypeName | null): any {
    return value as any;
}

export function BrandArchetypeNameToJSONTyped(value: any, ignoreDiscriminator: boolean): BrandArchetypeName {
    return value as BrandArchetypeName;
}